trade-show-scraping/requirements.txt
requests
beautifulsoup4
pandas
openpyxl
lxml
playwright
python-dotenv
